refactor(core): Move model ID normalization to shared utils
Keep canonical model ID resolution independent from token counting so all runtime consumers share the same helper.
