+++ packages/@n8n/agents/src/runtime/loop/__tests__/runtime-context.test.ts
@@ -1,5 +1,5 @@
 import type { ModelConfig } from '../../../types/sdk/agent';
-import { getModelIdString } from '../../model/model-token-counter';
+import { getModelIdString } from '../../../utils/model';
 
 describe('getModelIdString', () => {
 	it('returns a plain string model id unchanged', () => {
