+++ packages/@n8n/agents/src/runtime/model/model-token-counter.ts
@@ -1,18 +1,7 @@
 import type { ModelConfig } from '../../types/sdk/agent';
+import { getModelIdString } from '../../utils/model';
 
 export type TokenCounter = (text: string) => number | Promise<number>;
-
-/** Resolve a model config to its canonical `provider/model` id string. */
-export function getModelIdString(model: ModelConfig): string {
-	if (typeof model === 'string') return model;
-	if ('id' in model && typeof model.id === 'string') return model.id;
-	if ('modelId' in model && typeof model.modelId === 'string') {
-		const rawProvider = 'provider' in model ? String(model.provider) : 'unknown';
-		const provider = rawProvider.split('.')[0];
-		return `${provider}/${model.modelId}`;
-	}
-	return 'unknown';
-}
 
 function createTokenCounter(encoding: 'cl100k_base' | 'o200k_base'): TokenCounter {
 	return async (text) => {
